more fun with build tags
